pyproject.toml
[project]
name = "pineclaw"
version = "0.1.0"
description = "Lightweight AI Agent with Feishu as collaborative space"
requires-python = ">=3.12"
dependencies = [
    "lark-oapi>=1.5.3",
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "aiosqlite",
    "pydantic-settings",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
